+++ src/service/worker/runtime/statusTrack.ts
@@ -24,6 +24,7 @@ export const isStatusTrackBroadcast = (message: any): message is typeof BROADCAS
   message === BROADCAST_STATUS_TRACK
 
 export const statusTrackHandler = async (ctx: ServiceContext) => {
+  ctx.tracing?.currentSpan.setOperationName('builtin:status-track')
   if (!LINKED) {
     process.send?.(BROADCAST_STATUS_TRACK)
   }