+++ src/tracing/errorReporting/ErrorKinds.ts
@@ -1,0 +1,4 @@
+export const enum ErrorKinds {
+  REQUEST_ERROR = 'RequestError',
+  GENERIC_ERROR = 'GenericError',
+}